+++ lib/objects/flying_attack_angle_object.dart
@@ -13,7 +13,7 @@ import 'package:flutter/widgets.dart';
 
 class FlyingAttackAngleObject extends AnimatedObject
     with ObjectCollision, Lighting {
-  final int id;
+  final dynamic id;
   final FlameAnimation.Animation flyAnimation;
   final FlameAnimation.Animation destroyAnimation;
   final double radAngle;