+++ lib/util/mixins/attackable.dart
@@ -4,6 +4,6 @@ import 'package:flutter/widgets.dart';
 mixin Attackable on GameComponent {
   bool isAttackablePlayer = false;
   bool isAttackableEnemy = false;
-  void receiveDamage(double damage, int from);
+  void receiveDamage(double damage, dynamic from);
   Rect rectAttackable();
 }