+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.7.3]
 * hotfix: notify finish in `AnimatedObjectOnce`.
+* update `id` to dynamic in `receiveDamage`  and `JoystickAction`.
 
 ## [0.7.2]
 * hotfix render last frame in `AnimatedObjectOnce`.